fix: исправить загрузку инфо ордер часть 4 + сумма
